CoastCapitalPersonalAssistant/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.3.0
anthropic==0.40.0
feedparser==6.0.11
requests==2.31.0
caldav==1.3.9
python-dotenv==1.0.0
icalendar==5.0.12
pytz==2024.1
gunicorn==21.2.0
Jinja2==3.1.4
python-dateutil==2.9.0

# Testing
pytest>=8.0
pytest-mock>=3.12
freezegun>=1.4
prometheus_client==0.21.0
